Move helper fields/methods to base test if need different area flavor
