+++ richtextfx/src/integrationTest/java/org/fxmisc/richtext/InlineCssTextAreaAppTest.java
@@ -1,36 +1,19 @@
 package org.fxmisc.richtext;
 
 import javafx.geometry.Pos;
-import javafx.scene.Node;
 import javafx.scene.Scene;
 import javafx.scene.control.ContextMenu;
 import javafx.scene.control.MenuItem;
 import javafx.scene.input.MouseButton;
 import javafx.stage.Stage;
-import javafx.stage.Window;
 import org.testfx.api.FxRobotInterface;
-import org.testfx.framework.junit.ApplicationTest;
 import org.testfx.service.query.PointQuery;
-
-import static org.junit.Assume.assumeTrue;
 
 /**
  * TestFX tests should subclass this if it needs to run tests on a simple area. Any view-related API needs to be
  * wrapped in a {@link #interact(Runnable)} call, but model API does not need to be wrapped in it.
  */
-public class InlineCssTextAreaAppTest extends ApplicationTest {
-
-    static {
-        String osName = System.getProperty("os.name").toLowerCase();
-
-        isWindows = osName.startsWith("win");
-        isMac = osName.startsWith("mac");
-        isLinux = osName.startsWith("linux");
-    }
-
-    public static final boolean isWindows;
-    public static final boolean isMac;
-    public static final boolean isLinux;
+public class InlineCssTextAreaAppTest extends RichTextFXTestBase {
 
     public Stage stage;
     public Scene scene;
@@ -56,18 +39,6 @@ public class InlineCssTextAreaAppTest extends ApplicationTest {
 
         // so tests don't need to do this themselves
         area.requestFocus();
-    }
-
-    public final PointQuery position(Scene scene, Pos pos, double xOffset, double yOffset) {
-        return point(scene).atPosition(pos).atOffset(xOffset, yOffset);
-    }
-
-    public final PointQuery position(Window window, Pos pos, double xOffset, double yOffset) {
-        return point(window).atPosition(pos).atOffset(xOffset, yOffset);
-    }
-
-    public final PointQuery position(Node node, Pos pos, double xOffset, double yOffset) {
-        return point(node).atPosition(pos).atOffset(xOffset, yOffset);
     }
 
     public final PointQuery position(Pos pos, double xOffset, double yOffset) {
@@ -98,51 +69,4 @@ public class InlineCssTextAreaAppTest extends ApplicationTest {
         return clickOnFirstLine(MouseButton.SECONDARY);
     }
 
-    /**
-     * If not on Windows environment, calling this in @Before method will skip the entire test suite whereas calling
-     * this in @Test will skip just that test method
-     */
-    public final void run_only_on_windows() {
-        assumeTrue(isWindows);
-    }
-
-    /**
-     * If not on Linux environment, calling this in @Before method will skip the entire test suite whereas calling
-     * this in @Test will skip just that test method
-     */
-    public final void run_only_on_linux() {
-        assumeTrue(isLinux);
-    }
-
-    /**
-     * If not on Mac environment, calling this in @Before method will skip the entire test suite whereas calling
-     * this in @Test will skip just that test method
-     */
-    public final void run_only_on_mac() {
-        assumeTrue(isMac);
-    }
-
-    /**
-     * If on Windows environment, calling this in @Before method will skip the entire test suite whereas calling
-     * this in @Test will skip just that test method
-     */
-    public final void skip_if_on_windows() {
-        assumeTrue(!isWindows);
-    }
-
-    /**
-     * If on Linux environment, calling this in @Before method will skip the entire test suite whereas calling
-     * this in @Test will skip just that test method
-     */
-    public final void skip_if_on_linux() {
-        assumeTrue(!isLinux);
-    }
-
-    /**
-     * If on Mac environment, calling this in @Before method will skip the entire test suite whereas calling
-     * this in @Test will skip just that test method
-     */
-    public final void skip_if_on_mac() {
-        assumeTrue(!isMac);
-    }
 }
